Initialize single instance of workManager
Made changes to how the workManager starts a work to
make sure the workManager is started once and also
the existing work is kept and finished before
another work request can take place.

Set the style of the notification to enable users
to expand and read the full notification body.
Also changed the destination of the notification to
the homeFragment and did few code refactoring and
cleanups.