Make app single page application
Search results are now displayed on the same fragment
with few animations to remove the search button and
the search box and replace them with two new ones
which can reverse the effect that took place.

Didn't delete the search fragment. Left as the
destination for the notification's pending intent.