Add bottom navigation view
Added a bottom navigation view and added a
fragment that lists all the weather data
stored in the database and display them in
a recycler view.

App now displays a snack bar like before when
user enters an incorrect city name.

Update ReadMe with screenshots of current state
of the app.