Migrate and update database
Added requirements to migrate the database to new version
and also add a default value for the isFavourite column for
data already in the database.

Also removed performance of database queries on main thread
and added a custom threading class to the project, which i
will later refactor the code to use the custom threading
class.